feat(rspack)!: remove application generator in favour of framework-specific options (#32755)
Remove the Rspack Application Generator in favour of generators from
`@nx/react`, `@nx/angular` and other plugins with app generators.

Resolves NXC-3109
